owner_bot/pyproject.toml
[project]
name = "owner-bot"
version = "0.1.0"
description = "Telegram bot for store owner to manage products via Google Sheets"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.6.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
    "google-api-python-client>=2.139",
    "google-auth>=2.33",
    "Pillow>=10.0",
    "sentry-sdk>=2.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
